pyproject.toml
[build-system]
requires = [
    "setuptools>=65",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "naja_atrar"
description = "This is a simple http server, use MVC like design."
readme = "README.md"
authors = [
    { name = "keijack", email = "keijack.wu@gmail.com" }
]
requires-python = ">=3.7"
keywords = ["http-server", "websocket", "http", "web", "web-server"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["websocket-client"]

[tool.setuptools.packages.find]
include=["naja_atrar"]

[tool.setuptools.dynamic]
version = {attr = "naja_atrar.version"}

[project.urls]
homepage = "https://github.com/keijack/python-simple-http-server"
repository = "https://github.com/keijack/python-simple-http-server"